backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
motor
dnspython
pydantic
google-genai
